+++ modules/uu/tarificator/AccountLogResourceTarificator.php
@@ -55,8 +55,21 @@ class AccountLogResourceTarificator extends Tarificator
         $dateTimeOffsetParams = new DateTimeOffsetParams($this);
         $utcDateTime = $dateTimeOffsetParams->getCurrentDateTime();
 
+        $fromId = $toId = null;
+
+        // распаралелливание обработки
+        if (isset($_SERVER['argv']) && count($_SERVER['argv']) == 4 && $_SERVER['argv'][1] == 'ubiller/resource') {
+
+            $fromId = (int)$_SERVER['argv'][2];
+            $toId = (int)$_SERVER['argv'][3];
+
+            if (!$fromId || !$toId || $fromId > $toId) {
+                throw new \InvalidArgumentException('Неверные аргументы');
+            }
+        }
+
         // в целях оптимизации удалить старые данные
-        if (!$accountTariffId) {
+        if (!$accountTariffId && !$fromId && !$toId) { // не удалять, если расчет идет по услуге или идет параллельная обработка
             AccountLogResource::deleteAll(['<', 'date_from', $minLogDatetime->format(DateTimeZoneHelper::DATE_FORMAT)], [], 'id ASC');
         }
 
@@ -75,18 +88,7 @@ class AccountLogResourceTarificator extends Tarificator
                 ['<', 'account_log_resource_utc', $utcDateTime->format(DateTimeZoneHelper::DATE_FORMAT)] // или списаны давно
             ]);
 
-        // распаралелливание обработки
-        if (isset($_SERVER['argv']) && count($_SERVER['argv']) == 4 && $_SERVER['argv'][1] == 'ubiller/resource') {
-
-            $fromId = (int)$_SERVER['argv'][2];
-            $toId = (int)$_SERVER['argv'][3];
-
-            if (!$fromId || !$toId || $fromId > $toId) {
-                throw new \InvalidArgumentException('Неверные аргументы');
-            }
-
-            $accountTariffQuery->andWhere(['between', 'id', $fromId, $toId]);
-        }
+        $fromId && $toId && $accountTariffQuery->andWhere(['between', 'id', $fromId, $toId]);
 
         $accountTariffQuery
             ->with('clientAccount')
